Added no-op definitions for various functions.
icc was complaining about (declared) functions that were not defined but
that appeared in (un-executed) blocks of code. No clean solution using
inheritance presented itself. Instead, sparing use of preprocessor
directives checking the values of macros was employed, and this only in
two header files (coordinates.hpp and eos.hpp). The idea is to provide
no-op definitions for functions in cases where the configure options
would result in choosing .cpp files that do not define them.