*** READ THIS *** This commit contains many changes and is incompatible with previous versions. First, now we changed block boundary routines. Previously the boundary conditions are transferred in the x1 direction first, then x2, and finally x3 to fill the edge and corner ghost zones. In this version, now the edge and corner ghost zones are filled by communicating with diagonal neighbor blocks directly. This needs more MPI messages but all the communications are independent and the code is actually simpler. Please contact me if you have any question about this new boundary implementation.
Moreover, physical boundary condition functions now take the loop limits (is, ie, etc.) as parameters so that we can use these boundary functions for those edge/corner boundaries as well as for SMR prolongation. All the pgen files are already corrected, so please follow them to implement your own boundary conditions.

These changes are made so as to make SMR/AMR easier.

This is a birthday present for my self.