Fixed extremely subtle bug in GRMHD EMF calculation:
The 1D array of cell-centered metric values was not being populated in
the innermost and outermost x1-direction cells when calculating the
cell-centered electric field in the 2-direction. As a result, the
edge-centered electric fields at the inner and outer 1-faces of the
active zone were incorrect.

Amazingly, the error sat undetected for a year and a half. It only
appeared in a particularly demanding problem, using HLLD (diffusion
eliminates it), and even then only showing itself over the course of
many thousands of timesteps.

The 1D array in question is earlier populated for calculating the other
components of the E-field. The values being used erroneously were thus
those of the outermost x3-coordinate in the active zone and the
x2-coordinate just beyond the active zone. x3 is ignorable in all
current coordinate systems, and blocks that only have narrow extents in
x2 mitigate the error. Moreover, across block boundaries the two blocks
always agree on the electric field, wrong though they may be. The
cell-centered fields have only one purpose: to be used in calculating
electric field gradients that are upwinded and differenced to create the
effective Laplacian term in the Gardiner & Stone algorithm. Thus in some
cases upwinding eliminated the error altogether, and even in other cases
the error was in the amount of diffusion being applied.

Unsuccessful attempts to defeat and/or locate the bug included:
  - Using D and E+D rather than D and E as conserved variables (the
    former being what Harm does). This avoids catastrophic cancellation
    issues in certain parts of parameter space.
  - Restructuring the communication protocol so that variable inversion
    is done first and both primitives and conserved values are exchanged
    (to prevent roundoff errors combined with inversion failure from
    leading to wildly different results on two representations of the
    same cell).
  - Restarting a simulation just as its values were about to begin the
    long trend toward being completely wrong, but with the domain
    decomposition redone. This required manipulating the restart files
    on a byte-by-byte level in a very architecture- and
    compiler-dependent way.
  - Turning off all optimization, and even adding "-fp-model precise
    -fp-model source" flags to eliminate (the surprisingly abundant)
    cases where differences in domain decomposition led to
    floating-point discrepancies in the least significant few bits.
    These discrepancies were enough to mask the real problem in many
    cases.