Added arbitrary coordinates ability for GR:
This required some restructuring of the coordinates and pgen files. Some
helper functions for GR pgens (TransformVector..., Get...Coordinates,
etc.) were moved into the pgens themselves. This avoids issues with the
functions needed to be enrolled in cases where they are not needed.

In order to use the new capability, the code should be configured with
"--coord gr_user". The user must then enroll a metric function via
Mesh::EnrollUserMetric(), much like enrolling other user-defined
functions. The user is responsible for returning the covariant *and*
contravariant components of the metric (no automatic numerical matrix
inversion provided), as well as the spatial derivatives (no numerical
differencing provided), as a function of position.

Because this will require modifying a pgen file, the decision was made
to not expect existing pgens to work out of the box with some new
coordinates.

Internally, this new coordinates class works by storing everything in 3D
arrays, which are then accessed by the standard interfacing functions
(CellVolume(), Face1Area(), etc.). This will obviously use more memory.
Surprisingly, a preliminary test showed a slowdown of only 4% for a
torus using user-defined Kerr-Schild coordinates instead of the built-in
ones.