Improved Python HDF5 reader:
Default refinement level to output is now maximum level in file.

If the requested level will cause slow restriction or unnecessary
prolongation, a warning will be issued.

If any of the requested quantities are not available in the file, the
error message will list all valid possibilities.

Cell data arrays (not coordinates) are now returned as 32-bit floats
rather than 64-bit. This can be overridden with a new optional
parameter.