Fixed summation outputs with HDF5:
Previously, the data was summed in the given dimension, but the output
had the full number of dimensions, filling most of the cells with 0s.
Now summation really does produce smaller files, as intended.

The Python reader has also been updated to handle summed data. As with
all other forms of data, a uniform grid is returned. It is up to the
user to use this data appropriately, especially in the tricky case of
mesh refinement where different parts of the grid incorporate data from
different extents in the summed direction.