Improve symmetry-preservation of HLLD floating point operations
The big issue was the use of L-state for intermediate star
pressure that relied on exact cancellation of rounded FP terms.

Now, 90 degree rotational symmetry is preserved for hydro O-T
problem, like it was for HLLE. MHD result is still asymmetric
due to floating point associaitivity round off asymmetry in
upwinding of emf, etc.