Significantly changed GRMHD inversion iteration:
Now the i-loop is fissioned into 5 parts. The first pass applies the
iterative algorithm exactly 3 times, which in practice seems to converge
for the vast majority of cells. The next pass takes any non-converged
cells and applies the iteration algorithm 12 times. The last 3 passes
apply floors and check for consistency.