Changed CenterWidth functions in GR:
Formerly, these returned (possibly lower bounds on) proper distances
across cells. That is, W^i >= \int \sqrt{g_{ii}} dx^i. This was used in
two places: smoothing the upwinding of electric fields and calculating
the timestep.

The upwinding weight is of the form
    1024 * (\Delta t) * (\rho u^i) / (W^i * (\rho_L + \rho_R)).
With W^i having units of length, this formula has units of coordinate
per unit time, which is not dimensionless as it probably should be.

The timestep was being calculated as proper width, which corresponds to
the light crossing time. It was then changed (571ca08) to divide by the
values returned by FastMagnetosonicSpeedsGR() and SoundSpeedsGR().
However, these functions return speeds in units of coordinate per unit
coordinate time, so the result had the wrong units.

This commit makes all GR coordinates have the default behavior of
defining cell widths as changes in coordinate, whether this has units of
proper distance or not. This fixes both problems above.