Added ghost-zone awareness to HDF5 reader:
The reader now takes an optional argument, num_ghost, defaulting to 0.
If correctly provided (i.e. matching the number of ghost cells output in
the file with the ghost_zones flag), the output array will be the active
zone plus num_ghost cells on all nontrivial borders.

Note a number of other reader functions are incompatible with this
feature, and so errors will be raised if they are combined. In
particular, only uniform grids with essentially no refinement will work.

The reader does not attempt to infer num_ghost from the HDF5 file,
leaving this up to the user. Indeed, this is a very tricky thing to do
given the file data. As a result, plotting scripts that call the reader
will not presently work with ghost zones.

Also note possible unexpected behavior: When ghost_zones is set, the
HDF5 file contains not just the ghost zones at the edges of the domain,
but also all internal ghost zones between blocks. This is because the
ghost cells are generally only used for debugging. The reader makes no
attempt to convey multiple values existing at the same physical
locations. In fact, for cells that exist as both active and ghost cells,
the returned value may come from the ghost cell.

This addresses #184.