Use custom string processing instead of 'choices' list for --cxx flag
Now, 'gcc', 'clang', 'icc' will be mapped to the matching C++ front
ends, regardless of future changes to cxx_choices list. And only the C++
front ends will be shown as the "main" valid choices for --cxx, to avoid
confusion and clutter.

The optional 'choices' keyword argument for argparse.add_argument() is
inflexible for handling regex processing of string inputs, so now this
built-in check is disabled.

'metavar' was ultimately used to preserve the same ./configure.py -h
behavior (instead of showing "--cxx CXX' without the list of "main" C++
compiler choices)