New BoundaryVariable objects must always add pointer to bvars vector
This is currently the only way to ensure that copying data between
MeshBlocks located on the same MPI rank / serial process is
successful (copies to the BoundaryVariable object occupying the same
vector index and sets the BoundaryStatus on the target).

Otherwise, it is unnecessary to add FFT self-gravity's BoundaryVariable
object to the bvars vector; it is now automatically invoked in
BoundaryCommunication calls in Mesh::Initialize(), time_integrator.cpp
task list. Change this approach.