Create separate vector of refs to BoundaryVariable objects
This will allow a reference to the FFT self-gravity BoundaryVariable
object to remain in the original bvars vector (the "global set" of
BoundaryVariable objects) necessary for the current MeshBlock-MeshBlock
transfer on the same MPI rank, while also allowing it to opt-out of the
BoundaryValues::StartReceiving(phase), etc. BoundaryCommunication
wrapper function calls that occur in Mesh::Initialize() and
TimeIntegratorTaskList functions.

This provides another, higher level of switching than the recently-added
"enum class BoundaryCommSubset", which controls a sub-BoundaryVariable
level of granularity (i.e. whether or not the bd_var_ and/or
bd_val_flcor_ is involved in that particular StartReceiving() or
ClearBoundary() call in mesh initialization, etc.). Need to document and
define this convention.