Explicitly type cast std::complex* to fftw_complex* in library fn calls
http://www.fftw.org/fftw3_doc/Complex-numbers.html

C++ has its own complex<T> template class, defined in the standard
<complex> header file. Reportedly, the C++ standards committee has
recently agreed to mandate that the storage format used for this type be
binary-compatible with the C99 type, i.e. an array T[2] with consecutive
real [0] and imaginary [1] parts.
...
To the extent that this is true, if you have a variable
complex<double> *x, you can pass it directly to FFTW via
reinterpret_cast<fftw_complex*>(x).

------------

Might have issues with non-double precision compatibility?
E.g. configuring Athena++ with -float for single-precision, Real = float
std::complex<float>* are being typecast to fftw_complex*, which might be
assuming doubles.... If so, change std::complex<Real> to
std::complex<double>, or implement a more nuanced switch.

@changgoo:
In `plimton/`, there is `FFT_PRECISION` defined in `fft_2d.h` and
`fft_3d.h` it seems that the current FFT packages in `plimton/` does not
support the single precision.