Changed relativistic fast magnetosonic wavespeeds:
In both SR and GR, the magnetosonic equation is quartic. However, this
is only used for calculating wavespeeds as part of the Riemann solver.
There exists a quadratic whose roots give approximations to the
wavespeeds (overestimates, by at most a factor of 2). This has been used
successfully in GR MHD with nontransforming Riemann solvers. Now it is
used for all relativistic MHD wavespeed calculations. This method is
faster and more robust.