Fixed ContainVariable erroneously returning true when comparing strings of same length
I am directly committing this because it can affect many users and the fix itself is minor. This closes #416.

When string::find did not match, it returns std::npos but in fact it is internally -1. The last if statement in ContainVariable returned true when haystack and needle are of the same length. As a result, when 'prim' is specified in an output block, it also loaded 'cons' variables, causing the HDF5 writer to crash

Anyway, this is probably my last commit this year. I wish 2022 will be a great year for everyone.