Release Candidate of the Multigrid Poisson Solver
This commit contains the release candidate of the Multigrid Poisson Solver which is used in the method paper (Tomida & Stone 2023). The input parameters are slightly changed from the previous versions, and the convergence threshold is now defect itself, not multiplied by dx^2. Other than those, there are no major changes in the user interface. Additional test problems are provided.