Fixed GR coordinate speed of light when calculating timesteps:
The previous timestep calculation erroneously assumed light moves at a
coordinate speed of 1 in GR. In practice, this resulted in more
conservative timesteps than the CFL number would indicate, though for
user-defined metrics it could have lead to timesteps that were too
large. This is now fixed to be exact.

The code still assumes a signal speed of c rather than a magnetosonic
speed, and it still potentially (under)estimates the cell widths
themselves.

Fixes #545