Made HDF5 reader use double-precision coordinates:
Removed lines from code that forced coordinate arrays in particular to
match the chosen dtype (either user-specified or based on file
contents). This was fine in most cases, but resulted in catastrophic
failures when single precision was used to emulate double-precision
calculations in the code proper for spherical coordinates at very high
resolution.

Fixes #498