Fixing a linking error for chemistry with no network
The original code caused a linking error of outputs.cpp and history.cpp with icpc when compiled with -O0 because ChemNetwork::species_names (a static const member) was declared but not set. The error does not appear with -O3, possibly because it is optimized out. (I do not know why gcc has no problem with this.)

This fix adds none.cpp for an empty chemical network and initialize species_names.