Removed a degeneracy check from the HLLD and LHLLD Riemann Solvers.
It turned out that we should not remove the double-star states even when the normal magnetic field bx is very weak. If this branch is selected when bx is actually non-zero, it may return unphysical flux.

I do not think this has affected many people, as this is problematic only when the magnetic fields are very weak and the flow is not smooth. The probability that this branch is selected is very low in a realistic flow.

Thanks to Kazunari Iwasaki.