Remapping Bx in ghost cells of shear-periodic boundaries
In the present implementation, Bx in ghost cells of shear-periodic boundaries are calculated using the divB=0 constraint from remapped By and Bz. While this is mathematically correct, numerically, it may cause a large error in Bx particularly when By, Bz >> Bx. I revert it back to the old implementation which directly remaps Bx using the same algorithm. This violates the divB constraint in ghost cells but it should not matter.