Skip correction of EMF_z
The original implementation produces artifacts at shear-periodic boundaries when shearing box is used without orbital advection. I found that this is due to correction of the z-component of EMF. Conceptually this is demanded to conserve the total magnetic flux in the rotation (y) direction. However, this needs transformation of EMF_z, either adding or subtracting v_orb B_x. As EMF_z is defined at edge-center but B_x is defined on face-center, B_X must be interpolated, but naive arithmetic average underestimates the EMF. While higher-order interpolation may fix this issue for smooth flows, probably it does not work if B_x has discontinuity. Therefore, I concluded that the best solution is applying the EMF correction only to the y-component, leaving the z-component uncorrected - this is actually what is implemented in Athena (Stone & Gardiner 2010).

When shearing box and orbital advection are used together, the above transformation is not necessary at the shear periodic boundaries (instead it is processed by OA). In this case, we still can apply the EMF correction, and I think it is better to keep it.