Hot fix: load balancer
There was a case where the load balancer fails and assigns zero MeshBlock to the master rank (rank = 0). This occurs typically when the number of MeshBlocks is relatively small compared to the number of MPI ranks in use. With this fix, the balancer assigns at least one MeshBlock to every rank.

This closes Issue #688.